Remove unused queries from 969 prototype and rename client
